requirements.txt
openai==1.33.0
python-dotenv==1.0.1